requirements.txt
prompty[azure]
python-dotenv
numpy
pandas
azure-ai-evaluation
azure-identity
scikit-learn
matplotlib
seaborn
fastapi
uvicorn